requirements.txt
tweepy
pycountry
kafka-python
elasticsearch==7.11
pandas
flask
python-dotenv
